feat(lint-release-notes): allow for overriding of the github env variables when running
We need to be able to override the GitHub environment variables to allow for preview releases to run
on pull requests and through repository dispatch trigger events.

BREAKING CHANGE: `lint-release-notes` only runs on pull_request events.
When moving to v4, you will need to shift the usage to GHA that are
pull_request triggers.